% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{crop_nonvalid}
\alias{crop_nonvalid}
\title{crop_nonvalid}
\usage{
crop_nonvalid(.data, flag_nonvalid = T)
}
\arguments{
\item{.data}{a fitibble.}

\item{flag_nonvalid}{indicates whether the \code{is_nonvalid} flag should be added to \code{.data}}
}
\value{
a fitibble with nonvalid adherent wear entries, the rest of the entries are masked with an \code{NA}.
}
\description{
crop_nonvalid
}
\details{
Crops HR, steps and intensity minutes of nonvalid adherent wear.
}
