% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_adherent.R
\name{flag_adherent_hours_between}
\alias{flag_adherent_hours_between}
\title{flag_adherent_hours_between}
\usage{
flag_adherent_hours_between(time, hours_between = c(8, 20))
}
\arguments{
\item{time}{a numeric vector containing sequences of minute values in a day.}

\item{hours_between}{two-length numeric array taking lower/upper hour values between 0 and 24.}
}
\value{
boolean vector, adherent minutes encoded as T, and non-adherent encoded as F.
}
\description{
flag_adherent_hours_between
}
\details{
Flag adherent hours in the day. With this rule, the adherent time is bounded between lower/upper hour values as per the inclusion criteria of the study.
}
