% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_choi.R
\name{flag_choi_HR}
\alias{flag_choi_HR}
\title{Apply Choi et al nonwear algorithm to HR}
\usage{
flag_choi_HR(
  HR,
  min_period_len = 90,
  spike_tolerance = 2,
  min_window_len = 30,
  window_spike_tolerance = 0
)
}
\arguments{
\item{HR}{a numeric vector containing Fitbit HR for a subject.}

\item{min_period_len}{same as in flag_choi().}

\item{spike_tolerance}{same as in flag_choi().}

\item{min_window_len}{same as in flag_choi().}

\item{window_spike_tolerance}{same as in flag_choi().}
}
\value{
nonwear vector, non wear time encoded as T, and wear time encoded as F.
}
\description{
Apply Choi et al nonwear algorithm to HR
}
\details{
Modifying activity_threshold param is not optional because the presence of HR is binary.
}
\examples{
\dontrun{
minute_data \%>\%
  group_by(id) \%>\%
  mutate(HR = flag_choi_HR(HR))
}
}
