% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_choi.R
\name{flag_choi_steps}
\alias{flag_choi_steps}
\title{Apply Choi et al nonwear algorithm to steps}
\usage{
flag_choi_steps(
  steps,
  activity_threshold = 1,
  min_period_len = 90,
  spike_tolerance = 2,
  min_window_len = 30,
  window_spike_tolerance = 0
)
}
\arguments{
\item{steps}{a numeric vector containing Fitbit steps for a subject.}

\item{activity_threshold}{same as in flag_choi(), the difference is that the default activity threshold is specifically set for Fitbit steps to 1, however modifying the value is still optional.}

\item{min_period_len}{same as in flag_choi().}

\item{spike_tolerance}{same as in flag_choi().}

\item{min_window_len}{same as in flag_choi().}

\item{window_spike_tolerance}{same as in flag_choi().}
}
\value{
nonwear vector, non wear time encoded as T, and wear time encoded as F.
}
\description{
Apply Choi et al nonwear algorithm to steps
}
\examples{
\dontrun{
minute_data \%>\%
  group_by(id) \%>\%
  mutate(nonwear = flag_choi_steps(steps))
}
}
