% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_nonwear.R
\name{flag_nonwear}
\alias{flag_nonwear}
\title{flag_nonwear}
\usage{
flag_nonwear(
  HR,
  steps,
  nonwear_method = c("missing_HR", "missing_HR_zero_steps", "choi_HR", "choi_steps"),
  ...
)
}
\arguments{
\item{HR}{a numeric vector representing HR readings of a Fitbit.}

\item{steps}{a numeric vector representing step readings of a Fitbit. Depending on the nonwear method, either HR or steps or both are used.}

\item{nonwear_method}{a \code{character} entry, the name of one of the non wear method options: "missing_HR", "missing_HR_zero_steps", "choi_HR", or "choi_steps".}

\item{...}{additional parameters that go into the selected nonwear_method.}
}
\value{
nonwear vector, non wear time encoded as T, and wear time encoded as F.
}
\description{
flag_nonwear
}
\examples{
\dontrun{
minute_data \%>\%
  group_by(id) \%>\%
  mutate(
    missing_HR = flag_nonwear(HR, steps, "missing_HR"),
    missing_HR_zero_steps = flag_nonwear(HR, steps, "missing_HR_zero_steps"),
    choi_HR = flag_nonwear(HR, steps, "choi_HR"),
    choi_steps = flag_nonwear(HR, steps, "choi_steps")
  )
}
}
