% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_valid_day.R
\name{flag_valid_adherent_hours}
\alias{flag_valid_adherent_hours}
\title{flag_valid_adherent_hours}
\usage{
flag_valid_adherent_hours(is_wear, is_adherent, minimum_adherent_hours = 10)
}
\arguments{
\item{is_wear}{a logic vector with entries for every minute in day. \code{TRUE} values correspond to those minutes in the day that are considered to be wear time.}

\item{is_adherent}{a logic vector indicating minutes in a day that adhere to a study's inclusion criteria.}

\item{minimum_adherent_hours}{the minimum number of adherent hours required in a day to be considered valid.}
}
\value{
a single boolean indicating whether the minimum adherent hours criteria is satisfied.
}
\description{
flag_valid_adherent_hours
}
\details{
Flag valid days using minimum adherent hours criteria
}
