% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_valid_day.R
\name{flag_valid_day}
\alias{flag_valid_day}
\title{flag_valid_day}
\usage{
flag_valid_day(
  is_wear,
  is_adherent,
  steps,
  valid_day_method = c("valid_adherent_hours", "valid_step_count"),
  ...
)
}
\arguments{
\item{is_wear}{a logic vector with entries for every minute in day. \code{TRUE} values correspond to those minutes in the day that are considered to be wear time.}

\item{is_adherent}{a logic vector indicating minutes in a day that adhere to a study's inclusion criteria.}

\item{steps}{a numeric vector containing Fitbit minute step recordings in a day.}

\item{valid_day_method}{either "valid_adherent_hours" or "valid_step_count".}

\item{...}{additional parameters that go into the selected valid_day_method.}
}
\value{
a single boolean indicating whether the valid day criteria is satisfied.
}
\description{
flag_valid_day
}
\details{
Flag valid days using either minimum adherent hours or minimum step count criteria.
}
