% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_valid_day.R
\name{flag_valid_step_count}
\alias{flag_valid_step_count}
\title{flag_valid_step_count}
\usage{
flag_valid_step_count(is_wear, is_adherent, steps, minimum_step_count = 1000)
}
\arguments{
\item{is_wear}{a logic vector with entries for every minute in day. \code{TRUE} values correspond to those minutes in the day that are considered to be wear time.}

\item{is_adherent}{a logic vector indicating minutes in a day that adhere to a study's inclusion criteria.}

\item{steps}{a numeric vector containing Fitbit minute step recordings in a day.}

\item{minimum_step_count}{the minimum number of step counts required in a day to be considered valid.}
}
\value{
a single boolean indicating whether the minimum step counts criteria is satisfied.
}
\description{
flag_valid_step_count
}
\details{
Flag valid days using minimum step count criteria.
}
