% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{mask_fitibble}
\alias{mask_fitibble}
\title{mask_fitibble}
\usage{
mask_fitibble(.data, flag_valid = T)
}
\arguments{
\item{.data}{a fitibble.}

\item{flag_valid}{indicates whether the \code{is_valid} flag should be added to \code{.data}}
}
\value{
a fitibble with masked non-valid minutes.
}
\description{
mask_fitibble
}
\details{
Mask non-valid minute entries with \code{NA} entries.
}
