% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{mean_sed_bout_length}
\alias{mean_sed_bout_length}
\title{mean_sed_bout_length}
\usage{
mean_sed_bout_length(is_sedentary)
}
\arguments{
\item{is_sedentary}{a logic vector with entries for every minute in day. \code{TRUE} values correspond to those minutes in the day that are considered to be sedentary time.}
}
\value{
a single numeric value, the mean of sedentary bouts in a day.
}
\description{
mean_sed_bout_length
}
\details{
Calculates the mean sedentary bout in a day .
}
