% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{new_fitibble}
\alias{new_fitibble}
\title{new_fitibble}
\usage{
new_fitibble(
  x = tibble::tibble(),
  intensity_colname = character(),
  intensity_levels = character()
)
}
\arguments{
\item{x}{a tibble containing Fitbit HR and step minute readings for multiple subjects in a cohort.}

\item{intensity_colname}{a character entry indicating the name of the column that contains the intensity of each minute reading in \code{x}.}

\item{intensity_levels}{a named character vector indicating the labels and levels of the intensity categories.}
}
\value{
a fitibble object.
}
\description{
new_fitibble
}
