% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confidence_intervals.R
\name{plot_ci}
\alias{plot_ci}
\title{Plot confidence intervals}
\usage{
plot_ci(
  daily_summary,
  type = c("basic", "sedentary_behavior", "time_use"),
  critical_value = qnorm(p = 0.05/2, lower.tail = FALSE)
)
}
\arguments{
\item{daily_summary}{a tibble, ideally the output of \code{prep_daily_summary()}.}

\item{type}{the plot type, it can be \code{"basic"} or \code{"sedentary_behavior"} or \code{"time_use"}}

\item{critical_value}{the default value is a z-critical value with 95\%  confidence (assuming a normal distribution of the sample mean).}
}
\value{
A plot containing the confidence intervals of a physical activity PA descriptor in daily_summary. Each point range in the plot corresponds to a participant in a cohort.
}
\description{
Plot confidence intervals
}
