% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confidence_intervals.R
\name{plot_confidence_intervals}
\alias{plot_confidence_intervals}
\title{Plot confidence intervals}
\usage{
plot_confidence_intervals(
  daily_summary,
  mu_varname = "mu_HR",
  sd_varname = "sd_HR",
  type = c("continuous", "categorical"),
  critical_value = qnorm(p = 0.05/2, lower.tail = FALSE)
)
}
\arguments{
\item{daily_summary}{a tibble, ideally the output of \code{prep_daily_summary()}.}

\item{mu_varname}{a character entry indicating the column name that contains the sample means.}

\item{sd_varname}{a character entry indicating the column name that contains the sample standard deviations.}

\item{type}{the variable type, it can be \code{"continuous"} or \code{"categorical"}}

\item{critical_value}{the default value is a z-critical value with 95\%  confidence (assuming a normal distribution of the sample mean).}
}
\value{
A plot containing the confidence intervals of a PA descriptor in daily_summary. Each point range in the plot corresponds to a participant in a cohort.
}
\description{
Plot confidence intervals
}
