% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity_scatter.R
\name{plot_intensity_scatter}
\alias{plot_intensity_scatter}
\title{Intensity scatter plot}
\usage{
plot_intensity_scatter(minute_data, sample_size = 50000)
}
\arguments{
\item{minute_data}{a fitibble.}

\item{sample_size}{the number of points in the scatter plot}
}
\value{
A scatterplot with a two-dimension representation of the PA intensity levels. For each id a sample of minute HR and step readings is extracted. Colors represent PA intensity levels.
}
\description{
Intensity scatter plot
}
