% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intraday.R
\name{plot_intraday}
\alias{plot_intraday}
\title{Plot intraday}
\usage{
plot_intraday(minute_data, id, date, add_wear = T)
}
\arguments{
\item{minute_data}{a fitibble.}

\item{id}{the subject \code{id}.}

\item{date}{the date from which fitbit HR and step readings will be plotted.}

\item{add_wear}{default is \code{TRUE} to color the background of wear minutes in orange and nonwear minutes in gray.}
}
\value{
Intraday minute-level HR and step time series plot.
}
\description{
Plot intraday
}
