% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_use.R
\name{plot_time_use}
\alias{plot_time_use}
\title{Plot PA time-use}
\usage{
plot_time_use(minute_data, id)
}
\arguments{
\item{minute_data}{a fitibble.}

\item{id}{the subject \code{id}.}
}
\value{
A plot showing the time-use descriptors split in valid and non-valid days.
}
\description{
Plot PA time-use
}
