% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wear_heatmap.R
\name{plot_wear_heatmap}
\alias{plot_wear_heatmap}
\title{plot_wear_heatmap}
\usage{
plot_wear_heatmap(minute_data)
}
\arguments{
\item{minute_data}{a fitibble}
}
\value{
A heatmap indicating the wear proportion in each hour and weekday for each subject \code{id} in \code{minute_data}.
}
\description{
plot_wear_heatmap
}
