% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{prep_daily_data}
\alias{prep_daily_data}
\title{prep_daily_data}
\usage{
prep_daily_data(.data)
}
\arguments{
\item{.data}{a fitibble with readings of Fitbit data.}
}
\value{
a tibble containing daily statistics derived from minute-level data.
}
\description{
prep_daily_data
}
\details{
Prepare daily data.
}
