% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_minute_data.R
\name{prep_minute_data}
\alias{prep_minute_data}
\title{prep_minute_data}
\usage{
prep_minute_data(fitabase_files)
}
\arguments{
\item{fitabase_files}{the output list of \code{read_fitabase_files()}.}
}
\value{
A tibble containing \code{id} | \code{time} | \code{HR} | \code{steps} | \code{intensity}.
}
\description{
prep_minute_data
}
\details{
Prepare minute data
}
\examples{
\dontrun{
minute_data <- prepare_minute_data(fitabase_files)
minute_data \%>\%
mutate(hour = hour(time),
       weekday = wday(as.Date(time), label = TRUE))
}
}
