% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{prep_patient_summary}
\alias{prep_patient_summary}
\title{prep_patient_summary}
\usage{
prep_patient_summary(.data)
}
\arguments{
\item{.data}{a fittible.}
}
\value{
a tibble containing summary statistics for each subject \code{id} in \code{data}.
}
\description{
prep_patient_summary
}
