% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{print.fitibble}
\alias{print.fitibble}
\title{print.fitibble}
\usage{
\method{print}{fitibble}(x)
}
\arguments{
\item{x}{a fitibble.}
}
\value{
print fitibble content and arguments.
}
\description{
print.fitibble
}
