% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fitabase_files.R
\name{read_fitabase_files}
\alias{read_fitabase_files}
\title{Read Fitabase files}
\usage{
read_fitabase_files(zip_path)
}
\arguments{
\item{zip_path}{a path to a zip file that contains minute HR, steps and intensity files extracted from Fitabase. Filenames suffix contain '_heartrate_1min', '_minuteStepsNarrow', or '_minuteIntensitiesNarrow' followed by 'yyyymmdd_yyyymmdd.csv'.}
}
\value{
A list that contains:
\itemize{
\item \code{files}: a tibble that describes the files in zip path. It contains \code{filename} | \code{filetype} (HR, steps or intensity) | \code{lines} (number of lines in the file) | \code{label} (fitabase participant id label) | \code{id} (a newly assigned id) | \code{is_valid_file_length} (minimum file length check).
\item \code{raw_HR_list}, \code{raw_steps_list}, \code{raw_intensity_list}: lists of tibbles that contain raw minute HR, steps or intensity data for each fitabase participant.
\item \code{time_period}: a tibble that summarizes the minimum and maximum time recorded for each fitabase participant and file type (HR, steps and intensity). \code{is_valid_time_period} corresponds to the minimum time period check (at least one day of overlaping records from 0:00 to 23:59 hrs)
}
}
\description{
Read Fitabase files
}
\examples{
\dontrun{
fitabase_files <- read_fitabase_files(file.choose(new = FALSE))
fitabase_files$files
fitabase_files$time_period
lapply(fitabase_files$raw_HR_list, function(x) summary(x$HR))
lapply(fitabase_files$raw_steps_list, function(x) summary(x$steps))
lapply(fitabase_files$raw_intensity_list, function(x) summary(as.factor(x$intensity)))}
}
