% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{relevel_fitibble}
\alias{relevel_fitibble}
\title{relevel_fitibble}
\usage{
relevel_fitibble(.data, intensity_colname, intensity_levels)
}
\arguments{
\item{.data}{a fitibble.}

\item{intensity_colname}{a character entry indicating the name of the column that contains the intensity of each minute reading in \code{.data}.}

\item{intensity_levels}{a named character vector indicating the labels and levels of the intensity categories in \code{intensity_col}.}
}
\value{
a fitibble with redefined intensity attributes.
}
\description{
relevel_fitibble
}
\details{
Redefine \code{intensity_colname} and \code{intensity_levels} attributes in a fitibble.
}
