% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{validate_fitibble}
\alias{validate_fitibble}
\title{validate_fitibble}
\usage{
validate_fitibble(x)
}
\arguments{
\item{x}{a fitibble-like object.}
}
\value{
Checks a fitibble for internal consistency.
}
\description{
validate_fitibble
}
