% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{crop_valid}
\alias{crop_valid}
\title{crop_valid}
\usage{
crop_valid(.data, nonvalid = F, flag_valid = T, mask = F)
}
\arguments{
\item{.data}{a fitibble.}

\item{nonvalid}{determines if nonvalid wear should be cropped instead of valid wear}

\item{flag_valid}{indicates whether the \code{is_valid} flag should be added to \code{.data}.}

\item{mask}{indicates whether minutes that are not valid wear should be masked with an \code{NA} instead of filtered out. Only entries in HR, steps and intensity columns are masked.}
}
\value{
a fitibble of valid wear entries,
}
\description{
crop_valid
}
\details{
Crops minutes of valid wear in a fitibble.
}
