% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{crop_valid}
\alias{crop_valid}
\title{crop_valid}
\usage{
crop_valid(.data, flag_valid = T)
}
\arguments{
\item{.data}{a fitibble.}

\item{flag_valid}{indicates whether the \code{is_valid} flag should be added to \code{.data}}
}
\value{
a fitibble with valid adherent wear entries, the rest of the entries are masked with an \code{NA}.
}
\description{
crop_valid
}
\details{
Crops HR, steps and intensity minutes of valid adherent wear.
}
