% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{fitibble}
\alias{fitibble}
\title{fitibble}
\usage{
fitibble(
  .data,
  intensity_colname = "intensity",
  intensity_levels = c(sedentary = "0", light = "1", moderate = "2", active = "3"),
  nonwear_method = c("missing_HR", "missing_HR_zero_steps", "choi_HR", "choi_steps"),
  adherent_method = c("adherent_hours_between"),
  valid_day_method = c("valid_adherent_hours", "valid_step_count"),
  ...
)
}
\arguments{
\item{.data}{A tibble containing \code{id} | \code{time} | \code{HR} | \code{steps} | \code{intensity}. The intensity column can have another name, specified by \code{intensity_col}. It can be the output of \code{prepare_minute_data()} or a tibble or data.frame with columns  \code{id} (character)| \code{time} (POSIXct yyyy-mm-dd HH:MM:SS) |\code{HR} (numeric)|\code{steps} (numeric) and with minute-level values in each row. It is necessary that for each participant the minute sequence has no gaps and  missing records should be entered as \code{NA}.}

\item{intensity_colname}{a character entry indicating the name of the column that contains the intensity of each minute reading in \code{.data}.}

\item{intensity_levels}{a named character vector indicating the labels and levels of the intensity categories in \code{intensity_col}.}

\item{nonwear_method}{one of the following "missing_HR", "missing_HR_zero_steps", "choi_HR" or "choi_steps".}

\item{adherent_method}{one of the following "adherent_hours_between" (other adherence rule could be integrated into the package).}

\item{valid_day_method}{one of the following "valid_adherent_hours" or "valid_step_count".}

\item{...}{additional parameters that go into the nonwear, adherence or valid day methods.}
}
\value{
a fitibble, an object containing arguments that are useful for generating physical activity summaries from Fitbit data.
}
\description{
fitibble
}
\details{
Prepares a \code{fitibble} object.
}
