% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{prep_daily_data}
\alias{prep_daily_data}
\title{prep_daily_data}
\usage{
prep_daily_data(.data)
}
\arguments{
\item{.data}{a fitibble.}
}
\value{
a tibble containing daily statistics of valid wear for each date and subject \code{id} in \code{.data}.
}
\description{
prep_daily_data
}
\details{
Prepares daily data derived from valid wear in a fitibble.
}
