% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{prep_daily_summary}
\alias{prep_daily_summary}
\title{prep_daily_summary}
\usage{
prep_daily_summary(.data)
}
\arguments{
\item{.data}{a fitibble.}
}
\value{
a tibble containing summary statistics that describe daily patterns for each subject \code{id} in \code{.data}.
}
\description{
prep_daily_summary
}
