% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fitabase_files.R
\name{read_fitabase_files}
\alias{read_fitabase_files}
\title{Read Fitabase files}
\usage{
read_fitabase_files(zip_path)
}
\arguments{
\item{zip_path}{a path to a zip file that contains minute HR, steps and intensity files extracted from Fitabase. Filenames suffix contain '_heartrate_1min', '_minuteStepsNarrow', or '_minuteIntensitiesNarrow' followed by 'yyyymmdd_yyyymmdd.csv'.}
}
\value{
A list that contains: 1) Files: a tibble that describes the files in zip path. It contains | filename | filetype (HR, steps or intensity) | lines (number of lines in the file) | label (fitabase participant id label) | id (a newly assigned id) | min_lines (a file minimum length check), 2) raw_HR_list, 3) raw_steps_list, 4) raw_intensity_list: lists of tibbles that contain raw minute HR, steps or intensity data for each fitabase participant.
}
\description{
Read Fitabase files
}
\examples{
fitabase_files <- read_fitabase_files(file.choose(new = FALSE))
lapply(fitabase_files$raw_HR_list,
       function(x) summary(table(x$time)))
lapply(fitabase_files$raw_HR_list, function(x) summary(x$HR))
lapply(fitabase_files$raw_steps_list,
       function(x) summary(table(x$time)))
lapply(fitabase_files$raw_steps_list, function(x) summary(x$steps))
lapply(fitabase_files$raw_intensity_list,
       function(x) summary(table(x$time)))
lapply(fitabase_files$raw_intensity_list, function(x) summary(as.factor(x$intensity)))
}
