% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitibble.R
\name{relevel_fitibble}
\alias{relevel_fitibble}
\title{relevel_fitibble}
\usage{
relevel_fitibble(x, intensity_colname, intensity_levels)
}
\arguments{
\item{x}{a fitibble.}

\item{intensity_colname}{a character entry indicating the name of the column that contains the intensity of each minute reading in \code{x}.}

\item{intensity_levels}{a named character vector indicating the labels and levels of the intensity categories in \code{intensity_col}.}
}
\value{
releveled intensity in fitibble.
}
\description{
relevel_fitibble
}
